Refs #77. Checkpointing progress on live view.
